Separating arlula into its own file
