+++ arlulacore/__init__.py
@@ -1,4 +1,6 @@
-from .auth import Session
+from .auth import (
+    Session
+)
 
 from .archive import (
     CenterPoint,
@@ -16,23 +18,14 @@ from .orders import (
     OrdersAPI
 )
 
-class ArlulaAPI:
-    '''
-        Main class for Arlula API Calls. Contains an instance of the archive and orders APIs.
-    '''
+from .arlula import (
+    ArlulaAPI
+)
 
-    def __init__(self, session: Session):
-        self._archive = ArchiveAPI(session)
-        self._orders = OrdersAPI(session)
-    
-    def archiveAPI(self) -> ArchiveAPI:
-        '''
-            Returns the archive api instance, used for searching and ordering imagery.
-        '''
-        return self._archive
+from .common import (
+    ArlulaObject
+)
 
-    def ordersAPI(self) -> OrdersAPI:
-        '''
-            Returns the orders api instance, used for maintaining orders and getting resources.
-        '''
-        return self._orders
+from .exception import (
+    ArlulaSessionError
+)
